+++ packages/react/src/lib/godaddy/checkout-mutations.ts
@@ -62,6 +62,10 @@ export const CreateCheckoutSessionMutation = graphql(`
           enabled
             minimumOrderTotal
         }
+        gopay_override {
+          enabled
+          goPayAppId
+        }
       }
       shipping {
         originAddress {
@@ -147,6 +151,7 @@ export const CreateCheckoutSessionMutation = graphql(`
         operatingHours {
           pickupWindowInDays
           leadTime
+          pickupSlotInterval
           timeZone
           hours {
             monday {
@@ -190,6 +195,7 @@ export const CreateCheckoutSessionMutation = graphql(`
       defaultOperatingHours {
         pickupWindowInDays
         leadTime
+        pickupSlotInterval
         timeZone
         hours {
           monday {