+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,90 @@
 # nextjs
 
+## 0.1.40
+
+### Patch Changes
+
+- Updated dependencies [798d6d6]
+  - @godaddy/react@1.0.42
+
+## 0.1.39
+
+### Patch Changes
+
+- Updated dependencies [6f9b610]
+  - @godaddy/react@1.0.41
+
+## 0.1.38
+
+### Patch Changes
+
+- Updated dependencies [f43f2dd]
+  - @godaddy/react@1.0.40
+
+## 0.1.37
+
+### Patch Changes
+
+- Updated dependencies [a53681a]
+  - @godaddy/react@1.0.39
+
+## 0.1.36
+
+### Patch Changes
+
+- Updated dependencies [93b26ea]
+  - @godaddy/react@1.0.38
+
+## 0.1.35
+
+### Patch Changes
+
+- Updated dependencies [2f604ef]
+  - @godaddy/react@1.0.37
+
+## 0.1.34
+
+### Patch Changes
+
+- Updated dependencies [815fa6c]
+  - @godaddy/react@1.0.36
+
+## 0.1.33
+
+### Patch Changes
+
+- Updated dependencies [8d3f2c2]
+  - @godaddy/react@1.0.35
+
+## 0.1.32
+
+### Patch Changes
+
+- Updated dependencies [57d6115]
+  - @godaddy/react@1.0.34
+
+## 0.1.31
+
+### Patch Changes
+
+- Updated dependencies [353dc8b]
+  - @godaddy/localizations@1.0.9
+  - @godaddy/react@1.0.33
+
+## 0.1.30
+
+### Patch Changes
+
+- Updated dependencies [ed3cbef]
+  - @godaddy/react@1.0.32
+
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies [b3c0dad]
+  - @godaddy/react@1.0.31
+
 ## 0.1.28
 
 ### Patch Changes